+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Mark Container packet handled so it does not cause a print statement.
 * New Models for some blocks
-* Multi-Tool Assist render now shows the placed state of block
+* Multi-Tool Assist render now shows the placed state of block, and actually places that block
 
 -------------------------------------------------------------------
 **Version: 0.0.75 (BETA)**